Moving controller methods around to make them more separated based on domain.
