+++ backend/task-manager/src/main/java/com/example/task_manager/controller/AuthController.java
@@ -7,7 +7,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping("/auth-info")
+@RequestMapping("/api/auth")
 public class AuthController {
 
     private final AuthInfoService authInfoService;
