+++ backend/task-manager/src/main/java/com/example/task_manager/controller/TeamMemberController.java
@@ -2,10 +2,10 @@ package com.example.task_manager.controller;
 
 import com.example.task_manager.DTO.IsAssignedDTO;
 import com.example.task_manager.DTO.PasswordChangeRequestDTO;
-import com.example.task_manager.DTO.ResetPasswordRequestDTO;
 import com.example.task_manager.DTO.TaskDTO;
-import com.example.task_manager.DTO.TaskRequestDTO;
 import com.example.task_manager.DTO.TeamDTO;
+import com.example.task_manager.DTO.TeamMemberWithTeamLeadDTO;
+import com.example.task_manager.service.AdminService;
 import com.example.task_manager.service.TeamMemberService;
 
 import java.util.List;
@@ -14,52 +14,19 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping("/api/tasks")
+@RequestMapping("/api/members/actions")
 public class TeamMemberController {
 
     private final TeamMemberService teamMemberService;
+    private final AdminService adminService;
 
-    public TeamMemberController(TeamMemberService teamMemberService) {
+    public TeamMemberController(TeamMemberService teamMemberService, AdminService adminService) {
         this.teamMemberService = teamMemberService;
-    }
-
-    // Create a Task
-    @PostMapping
-    public ResponseEntity<?> createTask(@RequestBody TaskRequestDTO request) {
-        try {
-            TaskDTO task = teamMemberService.createTask(request);
-            return ResponseEntity.ok(task);
-        } catch (RuntimeException e) {
-            return ResponseEntity.badRequest().body(e.getMessage());
-        }
-    }
-
-    // Delete Task
-    @DeleteMapping("/{taskId}")
-    public ResponseEntity<?> deleteTask(@PathVariable int taskId) {
-        try {
-            teamMemberService.deleteTask(taskId);
-            return ResponseEntity.noContent().build();
-        } catch (Exception e) {
-            return ResponseEntity.badRequest().body(e.getMessage());
-        }
-    }
-
-    // Edit Task
-    @PutMapping("/{taskId}")
-    public ResponseEntity<?> editTask(@PathVariable int taskId,
-                                    @RequestBody TaskDTO taskDTO) {
-        try {
-            TaskDTO updatedTask = teamMemberService.editTask(taskId, taskDTO);
-            return ResponseEntity.ok(updatedTask);
-        } 
-        catch (RuntimeException e) {
-            return ResponseEntity.badRequest().body(e.getMessage());
-        }
+        this.adminService = adminService;
     }
 
     // Assign Member to Task
-    @PostMapping("/{taskId}/assign/{teamMemberId}")
+    @PostMapping("/task/{taskId}/assign/{teamMemberId}")
     public ResponseEntity<?> assignToTask(@PathVariable int taskId, @PathVariable int teamMemberId) {
         try {
             IsAssignedDTO assignedDTO = teamMemberService.assignToTask(taskId, teamMemberId);
@@ -70,7 +37,7 @@ public class TeamMemberController {
     }
 
     // Assign many members to a task
-    @PostMapping("/{taskId}/mass-assign")
+    @PostMapping("/task/{taskId}/mass-assign")
     public ResponseEntity<?> massAssignToTask(@PathVariable int taskId, @RequestBody List<Integer> teamMemberIds) {
         try {
             List<IsAssignedDTO> isAssignedDTOs = teamMemberService.massAssignToTask(taskId, teamMemberIds);
@@ -82,7 +49,7 @@ public class TeamMemberController {
     }
 
     // Change Password (Placeholder)
-    @PostMapping("/team-members/{teamMemberId}/change-password")
+    @PostMapping("/{teamMemberId}/change-password")
     public ResponseEntity<?> changePassword(@PathVariable int teamMemberId,
             @RequestBody PasswordChangeRequestDTO request) {
         try {
@@ -91,21 +58,9 @@ public class TeamMemberController {
         } catch (RuntimeException e) {
             return ResponseEntity.badRequest().body(e.getMessage());
         }
-    }
-
-    //Reset Password
-    @PostMapping("/team-members/{teamMemberId}/reset-password")
-    public ResponseEntity<?> resetPassword(@PathVariable int teamMemberId,
-            @RequestBody ResetPasswordRequestDTO request) {
-        try {
-            teamMemberService.resetPassword(teamMemberId, request.getNewPassword());
-            return ResponseEntity.noContent().build();
-        }
-        catch (RuntimeException e) {
-            return ResponseEntity.badRequest().body(e.getMessage());
-        }
-    }
+    }    
     
+    //Get all the teams that a team member is a member of
     @GetMapping("/{teamMemberId}/teams")
     public ResponseEntity<?> getTeamsForMember(@PathVariable int teamMemberId) {
         try {
@@ -116,11 +71,23 @@ public class TeamMemberController {
         }
     }
 
+    //get all assigned tasks that are assigned to a team member
     @GetMapping("/{teamMemberId}/tasks")
     public ResponseEntity<?> getAssignedTasks(@PathVariable int teamMemberId) {
         try {
             List<TaskDTO> tasks = teamMemberService.getAssignedTasks(teamMemberId);
             return ResponseEntity.ok(tasks);
+        } catch (RuntimeException e) {
+            return ResponseEntity.badRequest().body(e.getMessage());
+        }
+    }    
+
+    //get all team members stored in the database
+    @GetMapping("/team-members")
+    public ResponseEntity<?> getTeamMembers() {
+        try {
+            List<TeamMemberWithTeamLeadDTO> teamMembers = adminService.getAllTeamMembers();
+            return ResponseEntity.ok(teamMembers);
         } catch (RuntimeException e) {
             return ResponseEntity.badRequest().body(e.getMessage());
         }
